Survival Probabilities
The survival probabilities for mutant cells on BARW structures as a function of selective bias "s" and branching rate "b".
Total branch death cases are thrown out.
The N## in the filename denotes the lattice size N_0. The initial and final "s" and "b" are included as well.

Probabilities are collected at generation 1000 across 2000 different structures (none of which completely die).